Fix SkipScan distinct column identification
The SkipScan code assumed the first entry in PathKeys would match
the distinct column. This is not always true as postgres will remove
entries from PathKeys it considers constant leading to SkipScan
operating on the wrong column under those circumstances.
This did most likely not cause any wrong results as the other
constraints for SkipScan to apply still had to be satisfied but
resulted in very inefficient query execution for those affected
queries.
This patch refactors the SkipScan code to use the distinctClause
from the Query instead.

Fixes #3330